requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
pyyaml>=5.4.0
matplotlib>=3.4.0
scipy>=1.7.0
pyrealsense2>=2.50.0
transforms3d>=0.4.1
